- diminuindo quantidade de testes nos arquivos run.sh, run-is.sh, run-ep.sh - fixed: a verificação para a classe C estava retornando UNSUCCESSFUL, por causa da função getSeedFor

+++ ep/pthreads/ep-pthreads.c
@@ -158,8 +158,8 @@ pseudorandom(double *random_seed)
 double
 getSeedFor(double k)
 {
-    int             i;
-    int             j;
+    unsigned long   i;
+    unsigned long   j;
     double          b;
     double          t;
     int             m;
